patch: move geometry to public includes
