+++ libs/vxng/include/vxng/geometry.h
