+++ libs/vxng/include/vxng/vxng.h
@@ -3,6 +3,7 @@
 // This primary file exports all public headers
 
 #include "camera.h"       // IWYU pragma: export
+#include "geometry.h"     // IWYU pragma: export
 #include "orbit-camera.h" // IWYU pragma: export
 #include "renderer.h"     // IWYU pragma: export
 #include "scene.h"        // IWYU pragma: export
